Final commit with all test cases

+++ Urinals/src/Urinals.java
@@ -2,40 +2,52 @@ import java.util.ArrayList;
 import java.util.Scanner;
 import java.io.*;
 public class Urinals {
-    public static String optFile() throws IOException {
+    public static String optFile(String path) throws Exception {
 
 //        System.out.println("File Chosen");;
-        File f=new File("src/urinal.dat");
-        BufferedReader br= new BufferedReader(new FileReader(f));
-        String st;
-        ArrayList<Integer> count=new ArrayList<Integer>();
-        while((st=br.readLine())!=null){
-            count.add(numuri(st));
-        }
-        for(int i=0;i< count.size();i++){
-            System.out.println(count.get(i));
-        }
-        String s = "";
-        for(int i=0;i<count.size();i++){
-            s+=" "+count.get(i);
-        }
-        br.close();
-        File f2=new File("src/rule.txt");
-        int counter=1;
-        while(f2.exists()){
-            f2=new File("src/rule"+counter+".txt");
-            counter++;
-        }
+        try{
+            File f=new File(path);
+            if(!f.exists())
+                throw new Exception("File Not Found!!");
+            else if(f.length()==0)
+                throw new Exception("File Empty");
+            else{
+
+                BufferedReader br= new BufferedReader(new FileReader(f));
+                String st;
+                ArrayList<Integer> count=new ArrayList<Integer>();
+                while((st=br.readLine())!=null){
+                    count.add(numuri(st));
+                }
+                for(int i=0;i< count.size();i++){
+                    System.out.println(count.get(i));
+                }
+                String s = "";
+                for(int i=0;i<count.size();i++){
+                    s+=" "+count.get(i);
+                }
+                br.close();
+                File f2=new File("src/rule.txt");
+                int counter=1;
+                while(f2.exists()){
+                    f2=new File("src/rule"+counter+".txt");
+                    counter++;
+                }
 //        f2=File.createNewFile("rule"+counter+".txt");
 //        System.out.println(f2.getName());
-        FileWriter fw=new FileWriter(f2);
-        for(int i=0;i<count.size();i++){
-            fw.write(count.get(i)+"\n");
-        }
-        fw.close();
-        return s;
-//        return "File";
+                FileWriter fw=new FileWriter(f2);
+                for(int i=0;i<count.size();i++){
+                    fw.write(count.get(i)+"\n");
+                }
+                fw.close();
+                return s;
+            }
 
+        }
+        catch (Exception e){
+            System.out.println(e);
+        }
+        return null;
     }
     public static int numuri(String str){
         int count=0;
@@ -99,17 +111,17 @@ public class Urinals {
         String s= String.valueOf(n);
         return s;
     }
-    public static String optStr(String s) throws IOException {
+    public static String optStr(String s) throws Exception {
         if(s.equals("Input")){
             return optIn();
         }
         else if(s.equals("File")){
-            return optFile();
+            return optFile("src/urinal.dat");
         }
         return "Wrong Option Chosen";
     }
 
-    public static void main(String[] args) throws IOException {
+    public static void main(String[] args) throws Exception {
         System.out.println("Enter the input method(Input for keyboard and File for File input): ");
         Scanner sc=new Scanner(System.in);
         String s=sc.next();
